done third stage

+++ app/Http/Controllers/Admin/AdminBaseController.php
@@ -74,6 +74,16 @@ class AdminBaseController extends Controller
                 'icon' => 'icon-gift',
                 'hidden' => false,
             ],
+            'articles' => [
+                'key' => 'articles',
+                'icon' => 'icon-magazine',
+                'hidden' => false,
+            ],
+            'images' => [
+                'key' => 'images',
+                'icon' => 'icon-stack-picture',
+                'hidden' => false,
+            ],
         ];
         $this->breadcrumbs[] = $this->menu['home'];
     }
@@ -211,11 +221,13 @@ class AdminBaseController extends Controller
     {
         $this->validate($request, ['id' => 'required|integer|exists:'.$model->getTable().',id']);
         $table = $model->find($request->input('id'));
-        if (isset($table->image) && $table->image) {
+        if (isset($table->image) && $table->image && $table->image != 'images/placeholder.jpg') {
             $this->deleteFile($table->image);
         } elseif (isset($table->images)) {
             foreach ($table->images as $image) {
-                $this->deleteFile($image->image);
+                if ($image && $image != 'images/placeholder.jpg') {
+                    $this->deleteFile($image->image);
+                }
             }
         }
